+++ src/actions.js
@@ -2,4 +2,5 @@ export default {
   goto: 'goto',
   capture: 'capture',
   test: 'test',
+  wait: 'wait',
 };